Cleanup: Remove Python bindings for NaClDescs and IMC
These used to be used for testing, but the test that used them was
removed a while ago.

These bindings could still be useful in future for embedders of NaCl,
but the usefulness would be limited because the system Python often
isn't compiled for the same architecture as NaCl/sel_ldr.

Also, IMC and SRPC are deprecated.  We could remove the IMC operations
from the Python bindings and leave the rest, but it's simpler to remove
the Python bindings entirely.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3864
TEST=trybots
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/474673002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13619 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2