toolchain_build: Create 'work' package type and use it for target library compiler.
Create a new package type, called 'work', intended for non-memoized build
artifacts. Like a source package, it is not memoized and so will always run.
Unlike a source package, it will run even when the --sync flag is not used.

Use this mechanism to make the host toolchain to build target libraries with;
this replaces the earlier method of making a copy of the driver scripts and
using the driver config file to point to the LLVM and binutils output directly.
This simplifies the current build rules, and in a future CL will be used to
1) Easily allow the target libraries to be built on mac (currently the rpaths
   are not set up correctly for the gold plugin), and
2) Add the host libc++ to this target without modifying LLVM's rpath or using
   LD_LIBRARY_PATH

R=jvoung@chromium.org, dyen@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3623

Review URL: https://codereview.chromium.org/485823003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13632 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2