Fix regression in loading of mmap()'able nexes on Windows
r13580 caused a regression in loading an initial nexe on Windows, for
the case where the nexe descriptor is marked as loadable via mmap()
(with validation caching, etc.).

The problem is that NaClElfFileMapSegment() rounds the segment size up
to a multiple of 64k.  For some nexes, this means its Map() call tries
to map beyond the file's extent, which fails on Windows (but not Unix).
r13580 made this failure into a fatal error, whereas previously
NaClElfFileMapSegment() would fall back to non-mmap-based loading.

This failure occurred for nexes where:
 1) the code segment appears at the end of the file;
 2) the code segment isn't HLT-padded to a 64k-multiple size.

This includes QuickOffice's nexes.  Older versions of the PNaCl
toolchain (using the Gold linker) produce nexes with this layout.

I'm fixing this by restoring NaClHostDescMap()'s error-handling
behaviour from before r13580 (which that change removed), but only for
trusted mappings.  For untrusted memory mappings, we'll still handle
sandbox mmap holes robustly by aborting.

Restoring the old behaviour this way is intended to be a low-risk fix
that can be backported to branches.

In later changes, we should change NaClElfFileMapSegment() so that it
doesn't map beyond a file's extent.

I tested this manually with QuickOffice's point_x86_32.nexe.  With
"safe_for_mmap = 1" set in elf_util.c, sel_ldr prints the following,
indicating that the fallback works:

[...] MapViewOfFileEx failed at 0x039c0000, got 0x00000000, err 5
[...] NaClElfFileMapSegment: Could not make scratch mapping, falling back to reading
PpapiPluginStart: PPAPI hooks not found

I've also added a unit test for Map().  This refactors the test code to
split out MakeTempFileNaClDesc() for code reuse.

BUG=https://code.google.com/p/chromium/issues/detail?id=406632
TEST=run_service_runtime_tests + manual test of nexe load
R=ncbray@chromium.org

Review URL: https://codereview.chromium.org/464393004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13655 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2