Implemented the _exit() syscall in MinSFI.
The syscall allows to immediatelly terminate execution of the
sandbox. It is implemented with the C setjmp/longjmp exception
mechanism, and will jump back to the MinsfiInvokeSandbox function.

Calling exit() now becomes the only legal way of terminating
the sandbox, and returning from the entry function is considered
an error. This allows us to conform to the standard signature of
_start() which returns void.

BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= ./scons platform=x86_64 bitcode=1 minsfi=1 minsfi_tests
NOTRY=true

Review URL: https://codereview.chromium.org/569043002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13756 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2