Add CreateNaClDescWithFilePathMetadata fn.
This change adds a method to create a NaClDesc for a NaClHandle which has
file path information without requiring the emebedder to do some of the work
preparing the internals of the NaClDesc.

WithFilePathMetadata has been added to the end of this method name since ncbray
suggested that we may want to create validated NaClDescs using different
metadata types.

BUG=chromium:394130
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/556613002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13763 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2