Switch PNaCl lib, include, and translator directories for new layout
This makes the directory layout more like a standard gnu-style toolchain and more suitable for support of multiple architectures, which will be used in direct-to-nacl mode. For now, since the biased bitcode builds will have to be around until direct-to-nacl mode can replace them, they are included in the same scheme as if they were their own architectures.

For now the directories used by the translator (the "native libs" and translator nexes) have a similar layout to before except that they all move into the translator/ directory.

The changes to the bitcode lib builds should also make it easier to reuse them  for direct-to-nacl mode if desired.

R=jvoung@chromium.org, noelallen@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3946

Review URL: https://codereview.chromium.org/583783002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13764 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2