Remove archived_frontend test and use gcc build for FYI bots
The archived_frontend test has not passed in a long time, and there are
now several things broken about it.
1) The old frontend cannot build the tests because of header changes
in the toolchain.
2) SCons has knowledge of toolchain layout which has changed recently.
3) The old toolchain is downloaded differently than current toolchains.

We haven't been fixing such problems recently.
Similar problems have also occurred in the past so even if we were to
fix them we would still not be testing bitcode as old as we need; so
this test has outlived its usefulness. It's better to have the bots
be green most of the time so we see when other tests start failing.

There is work currently going on upstream on exhaustive bitcode tests
that we should contribute to instead.

Also switch the FYI bots to use the gcc build of the toolchain. The
teststuite has been failing since the libc++ switch because the native
binaries (which get built with LLVM's makefiles) require libc++ and
can't load it. Switching to the gcc build will ensure it stays working
and make the testsuite work again.

R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/594343002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13781 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2