command.py: Run SkipForIncrementalCommand rules if the working directory is empty except for a log file
SkipForIncrementalCommand rules are meant to be run only when doing a
clobber build or when necessary (e.g. when the working directory is
empty or nonexistent). A recent change caused the working directory to
be populated with a log file on cache hits, which defeated the "empty
or nonexisted logic" so extend that logic to include the log file.

R=dyen@chromium.org
BUG=

Review URL: https://codereview.chromium.org/617103004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13818 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2