toolchain_build: Always clean git working directories on bots
PNaCl's CheckoutGitBundleForTrybot is a mechanism to try LLVM changes
on the bots without committing them. It results in a dirty working
directory, which doesn't get properly cleaned up and causes problems
on later runs. This adds an extra clean step conditioned on being
a bot, and also only runs the checkout step on bots.

BUG=Red try runs
R=dyen@chromium.org

Review URL: https://codereview.chromium.org/616893003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13822 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2