allow 16 bit xadd (including atomic version) in ia32.
This was always allowed in 64 bit mode. So there is no reason to disallow it in ia32. the only reason it was disallowed is for compatibility with the old validator.

There is still a pending compatibility/bug reproduction in that xadd isn't marked as zero extending. However, leaving that issue alone in the particular change.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2981
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/687383002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14010 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2