Disable getrlimit/setrlimit in the LLVM translator, readv in gold.nexe.
Otherwise, with the new stubs in r14027 it thinks that getrlimit
and setrlimit are usable, and attempt to use struct rlimit, which
is not defined for NaCl newlib. This then fails to compile.

Similar for readv and gold. We cannot actually compile code that uses
readv because there is not sys/uio.h header in newlib.

BUG=none (red bot).
R=dschuff@chromium.org, sbc@chromium.org

Review URL: https://codereview.chromium.org/686273004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14043 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2