Run expand-varargs pass as well as expand-byval when using zerocost exceptions
On NaCl targets int64s and doubles use alignment of 8 instead of 4
(the equivalent of the -malign-double flag), whereas on x86-32 Linux they
have alignment of 4. However the alignment of double arguments passed on
the stack is still 4 in all cases. With le32 bitcode, the frontend leaves
va_arg calls as intrinsics which are expanded by the -expand-varargs pass
at link time. With x86-32-native bitcode, the frontend expands them itself.
When using --pnacl-allow-exceptions, the normal ABI simplification passes
are not run, but -expand-varargs must be run or this test will fail because
after LLVM 3.5 the backend will use 8-byte-aligned load to expand the va_arg
intrinsic.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3930
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/742333002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14144 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2