Nest the pnacl section of untrusted.gypi to the same level.
The pnacl sections of untrusted.gypi were nested one layer less deep
than other toolchains. This seems to result in >(_sources) being
expanded earlier for pnacl, resulting in misses sources in the
case of base_nacl.gyp in the chrome tree. This is likely because
this is one of the only places that we use target_defaults
to inject a non-variable 'sources' list.

Nesting to the same level seems to alleviate the issue.
(Hopefully giving us more correct dependencies.)

BUG=https://code.google.com/p/chromium/issues/detail?id=427427
TEST=local ninja check + trybots
R=mseaborn@chromium.org, derek@chromium.org

Review URL: https://codereview.chromium.org/684033002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14152 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2