Use static split module scheduling when using pnacl-translate.py
The chromium builders want to cache build results but the default translator behavior is to use dynamic scheduling for parallel translation, which reorders functions in the final nexe. This makes the build non-deterministic, which is nice to prevent predictable code generation in-browser (admitedly not a big security mitigation), but isn't nice for build caching. Static module scheduling will still use multiple cores to translate but is a tad slower.

R=jvoung@chromium.org, dschuff@chromium.org
BUG= https://code.google.com/p/chromium/issues/detail?id=429358
TEST= cd chromium/src && ninja -C out/Release ppapi_nacl_tests_pnacl_newlib_x64.nexe && mv out out.1 && gclient runhooks && ninja -C out/Release ppapi_nacl_tests_pnacl_newlib_x64.nexe && find ./ -name ppapi_nacl_tests_pnacl_newlib_x64.nexe | xargs -n1 md5sum

Review URL: https://codereview.chromium.org/776283004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14173 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2