Update upstream compiler_rt to HEAD and use official git mirror
This switches our local HEAD to a new copy of the compiler_rt history from
the official git mirror (previously we were using a locally-managed git-svn
branch because there was no official git mirror).
It also update compiler_rt to the upstream HEAD; this is mostly a no-op for
the builtin functions we use, but it brings in build system changes I will
use for nacl-clang.

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3946

Review URL: https://codereview.chromium.org/873003004