Remove file-type check from NaclHostDescOpen
When running sel_ldr with -a it is useful to be able to open POSIX
named pipes (aka FIFOs, created with the mkfifo shell command or
library function).  Having named pipes (created outside the sandbox,
e.g. by shell scripts on the host) available in the toolbox is helpful
for a new test harness scheme in glibc, for example.  Once S_ISFIFO is
admitted, the existing check does not actually exclude any real-world
cases at all.  The only other type that (in Linux) you can actually
open is S_IFDIR (directories).  NaclHostDescOpen is never reached for
directories except in a TOCTTOU race with the pre-open stat check done
in NaClSysOpen, and that does not seem useful to check for at this
layer.  (If we really cared, we would use a non-racy check instead at
the higher layer.)  The Windows implementation of NaClHostDescOpen
already has no such corresponding check.

BUG= none
TEST= trybots + local test on Linux after creating with mkfifo command
R=mseaborn@chromium.org, sbc@chromium.org

Review URL: https://codereview.chromium.org/956093003