Remember to plumb the -O flag from pnacl-translate to pnacl-sz.nexe.
Only flags in SZ_FLAGS_EXTRA are plumbed through.
The others are set to appropriate defaults at compile time
under -DPNACL_BROWSER_TRANSLATOR (e.g., the arch cannot be
set to cross-compile).

There's still a bug in the IRT interface vs what
pnacl-sz.nexe expects... The IRT generates an argc and
argv, but it doesn't fill argv[0] to the "program name" so
it takes the first flag as the program name and the flags
parser thinks there are no flags.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4091
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1049553002