Add O_EXCL support to sel_ldr open syscall
This is needed in particular for the newlib implementation
of mkstemp (and family) with relies on O_EXCL causing open
to fail for existing files.

Cleaned up the handling of open flags in the windows
implementation.  Previously it was handling truncates
after opening the file, but the windows flags seem to
handle all the cases we need without this extra complexity.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4017

Review URL: https://codereview.chromium.org/1030893004